+++ src/plane.h
@@ -48,7 +48,7 @@ public:
 	DECL_ATTR( Tone,      Tone&   )
 	DECL_ATTR( WaveAmp,   int     )
 	DECL_ATTR( WaveLen,   int     )
-	DECL_ATTR( WaveSpeed, int     )
+	DECL_ATTR( WaveSpeed, float   )
 	DECL_ATTR( WavePhase, float   )
 	DECL_ATTR( WaveMode,  int     )
 	DECL_ATTR( WaveSize,  int     )
